Implemented DB first

+++ PizzaBox.Domain/Abstracts/AComponents.cs
@@ -4,8 +4,7 @@ namespace PizzaBox.Domain.Abstracts
     {
         public string Name { get; set; }
 
-        public decimal Price { get; set; }
-
+        public double Price { get; set; }
 
     }
 }
